refactored theme web views
